+++ CHANGELOG.md
@@ -7,6 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
+## [Unresolved]
+### Added
+- Validators for specified types:
+  - Undefined
+  - NaN
+  - Array
+  - String
+  - Public key
+  - Secret key
+  - Balance key
+  - Int64 amount
+  - Operation subject
+  - Asset code
+  - Fee type
+  - Creator details
+- Check for spaces in `BaseOperation.isValidString`
+
 ## [1.6.0-x.0] - 2019-04-26
 ### Added
 - `DELETE` method for ApiCaller