+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
+## Unreleased
+###Added
+* Builders for `CancelPollOp`, `UpdatePollEndTime`
+* Unit tests for for `CancelPollOp`, `UpdatePollEndTime`
+* Integration tests for `CancelPollOp`, `UpdatePollEndTime`
+
 ## [1.6.0-x.0] - 2019-04-26
 ### Added
 - `DELETE` method for ApiCaller